requirements.txt
keras
numpy
matplotlib
sklearn
scipy
selenium
thinkx
progress
tensorflow
pillow
IPython
certifi
tflearn
pypng
flask